-Fix cast warnings -Fix signed/unsigned comparison warnings -Fix format warnings -Add MinGW 3.0 rc header
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@7261 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/llio.cpp
@@ -314,7 +314,7 @@ int LLIO_seek(
 			method = FILE_END;
 			break;
 		}
-		if (SetFilePointer((HANDLE) file_desc, offset, NULL, method) == -1) {
+		if (SetFilePointer((HANDLE) file_desc, offset, NULL, method) == (DWORD) -1) {
 			if (status_vector)
 				io_error(status_vector, "SetFilePointer", filename,
 						 isc_io_access_err);
